Compilation and tests pass with gcc6 (MLDB-1895)
