+++ ext/sqlite/sqlite3pp.h
@@ -131,7 +131,7 @@ namespace sqlite3pp
 
    protected:
     explicit statement(database& db, char const* stmt = 0);
-    ~statement();
+    ~statement() noexcept(false);
 
     int prepare_impl(char const* stmt);
     int finish_impl(sqlite3_stmt* stmt);
@@ -298,7 +298,7 @@ namespace sqlite3pp
   {
    public:
     explicit transaction(database& db, bool fcommit = false, bool freserve = false);
-    ~transaction();
+    ~transaction() noexcept(false);
 
     int commit();
     int rollback();
