+++ jml/utils/testing/live_counting_obj.h
@@ -38,7 +38,7 @@ struct Obj {
         magic = GOOD;
     }
 
-   ~Obj()
+    ~Obj() noexcept(false)
     {
         //cerr << "destroying at " << this << endl;
         atomic_add(destroyed, 1);
