+++ ml/jml/judy_array.h
@@ -536,7 +536,7 @@ public:
 
     judyl_typed & operator = (const judyl_typed & other)
     {
-        if (&other == this) return;
+        if (&other == this) return *this;
         clear();
         insert(const_pair_iterator(other.begin()),
                const_pair_iterator(other.end()));
@@ -753,7 +753,7 @@ public:
 
     judyl_typed & operator = (const judyl_typed & other)
     {
-        if (&other == this) return;
+        if (&other == this) return *this;
         clear();
         insert(const_pair_iterator(other.begin()),
                const_pair_iterator(other.end()));
