+++ ml/separation_stats.h
@@ -13,6 +13,7 @@
 #include "mldb/jml/math/xdiv.h"
 #include "mldb/jml/utils/rng.h"
 #include "mldb/ext/jsoncpp/json.h"
+#include <cmath>
 #include <boost/any.hpp>
 
 
@@ -98,10 +99,10 @@ struct BinaryStats {
     {
     	double num = truePositives() * trueNegatives() -
     				falsePositives() * falseNegatives();
-    	double den = sqrt( (truePositives()+falsePositives()) *
-                           (truePositives()+falseNegatives()) *
-                           (trueNegatives()+falsePositives()) *
-                           (trueNegatives()+falseNegatives()));
+    	double den = std::sqrt( (truePositives()+falsePositives()) *
+                                (truePositives()+falseNegatives()) *
+                                (trueNegatives()+falsePositives()) *
+                                (trueNegatives()+falseNegatives()));
     	return ML::xdiv(num, den);
     }
