+++ rest/testing/rest_collection_stress_test.cc
@@ -14,6 +14,7 @@
 #include <chrono>
 #include <thread>
 #include <functional>
+#include <numeric>
 
 #define BOOST_TEST_MAIN
 #define BOOST_TEST_DYN_LINK
