+++ sql/sql.mk
@@ -20,7 +20,8 @@ SQL_EXPRESSION_SOURCES := \
 	path.cc \
 	dataset_types.cc \
 	sql_expression_operations.cc \
-	eval_sql.cc
+	eval_sql.cc \
+	expression_value_conversions.cc
 
 # Unfortunately the S2 library needs you to mess with the include path as its includes
 # aren't prefixed.
