+++ utils/testing/compact_vector_test.cc
@@ -55,7 +55,7 @@ struct Obj {
         magic = GOOD;
     }
 
-   ~Obj()
+    ~Obj() noexcept(false)
     {
         //cerr << "destroying at " << this << endl;
         ++destroyed;
