+++ vfs_handlers/s3_handlers.cc
@@ -522,13 +522,6 @@ makeStreamingDownload(const std::string & uri)
     return make_pair(std::move(result), source.info());
 }
 
-std::pair<std::unique_ptr<std::streambuf>, FsObjectInfo>
-makeStreamingDownload(const std::string & bucket,
-                      const std::string & object)
-{
-    return makeStreamingDownload("s3://" + bucket + "/" + object);
-}
-
 
 /****************************************************************************/
 /* S3 UPLOADER                                                              */
@@ -799,16 +792,6 @@ makeStreamingUpload(const std::string & uri,
                  (StreamingUploadSource(uri, onException, metadata),
                   131072));
     return result;
-}
-
-std::unique_ptr<std::streambuf>
-makeStreamingUpload(const std::string & bucket,
-                    const std::string & object,
-                    const OnUriHandlerException & onException,
-                    const S3Api::ObjectMetadata & metadata)
-{
-    return makeStreamingUpload("s3://" + bucket + "/" + object,
-                               onException, metadata);
 }
 
 /** Register S3 with the filter streams API so that a filter_stream can be
